requirements.txt
numpy
pyzmq
opencv-python
imutils